refactor: `SummarySection` as a deeper abstraction
